restructure project to be Python module, add Dockerfile
