+++ .dockerignore
@@ -1,0 +1,3 @@
+Dockerfile
+.pixi/
+.github/
