+++ conda_metadata_app/pages/search_by_file_path_page.py
@@ -5,7 +5,7 @@ import requests
 import streamlit as st
 from streamlit_searchbox import st_searchbox
 
-from app import app_config
+from conda_metadata_app.app_config import AppConfig
 
 
 @st.cache_resource(ttl="15m", max_entries=100)
@@ -49,7 +49,7 @@ st.set_page_config(
     initial_sidebar_state="collapsed",
 )
 
-if not app_config().enable_filepath_search:
+if not AppConfig().enable_filepath_search:
     st.error("File path search is disabled in the app configuration.")
     st.stop()
